+++ src/Services/slideServices.js
